ファイルが既に存在していないことを確認する

+++ manager/uploadingfile.php
@@ -59,6 +59,13 @@ $saveFileBaseFilePath= $saveDirPath."/";
 $saveAppFilePath= $saveFileBaseFilePath .$saveFileName;
 
 
+//ファイルが存在しないことを確認する
+if(file_exists($saveAppFilePath)===true){
+    //すでにファイルがある場合
+    exit();
+}
+
+
 
 if($saveAppFilePath!=="") {
     if(!move_uploaded_file($uploadtedFile['tmp_name'], $saveAppFilePath)){
